# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ScriptForge API
  version: 1.0.0
  description: ScriptForge 前台 / 后台 REST API 契约
paths:
...
  /api/billing/recharge/orders/{order_no}/mock_pay/:
    post:
      operationId: billing_recharge_orders_mock_pay_create
      description: POST /api/billing/recharge/orders/<order_no>/mock_pay/
      parameters:
      - in: path
        name: order_no
        schema:
          type: string
        required: true
      tags:
      - billing
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: No response body
  /api/billing/recharge/packages/:
    get:
      operationId: billing_recharge_packages_retrieve
      description: GET /api/billing/recharge/packages/
      tags:
      - billing
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: No response body
...
  /api/creation/agents/catalog/:
    get:
      operationId: creation_agents_catalog_retrieve
      description: GET /api/creation/agents/catalog/ — Agent 体系 SSOT（registry v2）。
      tags:
      - creation
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          description: No response body
...
  /api/creation/fusion/catalog/:
    get:
      operationId: creation_fusion_catalog_retrieve
      description: GET /api/creation/fusion/catalog/ — 题材/平台/格式等枚举（SSOT，前端禁止硬编码）。
      tags:
      - creation
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          description: No response body
  /api/creation/fusion/nodes/:
    get:
      operationId: creation_fusion_nodes_retrieve
      description: GET /api/creation/fusion/nodes/ — 主链节点（Agent 元数据 + 币价）。
      tags:
      - creation
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          description: No response body
...
  /api/creation/projects/{project_id}/nodes/{node_index}/preview/:
    get:
      operationId: creation_projects_nodes_preview_retrieve
      description: GET /api/creation/projects/<project_id>/nodes/<node_index>/preview/
      parameters:
      - in: path
        name: node_index
        schema:
          type: integer
        required: true
      - in: path
        name: project_id
        schema:
          type: string
        required: true
      tags:
      - creation
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT